feat: Added instructions functions inicio/final
inicio() and final() checks if the PDI is pointing and the first ghost
node or the last ghost node respectively.

This function is useful to be used as a loop ender when we want to read all the elements of the list,
being the value that this function returns the limit of the loop.